Support Windows batch-mode
A better version of adding support for Windows batch-mode. On normal
setups no Stata code needs to changed. Test included.

+++ ado/parallel_source.sthlp
@@ -1601,6 +1601,7 @@ mata:
 *!{col 6}{bf:nclusters}{col 20}Number of clusters.
 *!{col 6}{bf:paralleldir}{col 20}Dir where the process should be running.
 *!{col 6}{bf:timeout}{col 20}Number of seconds to wait until stop the process for no conextion.
+*!{col 6}{bf:gateway_fname}{col 20}Name of file that a Cygwin process is listen to will execute from (Windows batch).
 *!{col 4}{bf:returns:}
 *!{col 6}{it:Number of clusters that stopped with an error.}
 *!{dup 78:{c -}}{asis}
@@ -1608,10 +1609,12 @@ real scalar parallel_run(
     string scalar parallelid, 
     |real scalar nclusters, 
     string scalar paralleldir,
-    real scalar timeout
+    real scalar timeout,
+    string scalar gateway_fname
     ) {
 
     real scalar fh, i
+    string scalar tmpdir_i
     
     // Setting default parameters
     if (nclusters == J(1,1,.)) nclusters = strtoreal(st_global("PLL_CLUSTERS"))
@@ -1654,22 +1657,34 @@ real scalar parallel_run(
         stata("winexec sh __pll"+parallelid+"_shell.sh")
     }
     else { // WINDOWS
-        
-        unlink("__pll"+parallelid+"_shell.bat")
-        fh = fopen("__pll"+parallelid+"_shell.bat","w", 1)
-        
-        // Writing file
-        for(i=1;i<=nclusters;i++) {
-            mkdir(tmpdir+"__pll"+parallelid+"_tmpdir"+strofreal(i, "%04.0f"),1)
-            fwrite(fh, "start /MIN /HIGH set TEMP="+tmpdir+"__pll"+parallelid+"_tmpdir"+strofreal(i,"%04.0f")+" ^& ")
-            fput(fh, paralleldir+`" /e /q do ""'+pwd()+"__pll"+parallelid+"_do"+strofreal(i,"%04.0f")+`".do"^&exit"')
+        if (c("mode")=="batch"){ //Execute commands via Cygwin process
+            if (gateway_fname == J(1,1,"")) gateway_fname = st_global("PLL_GATEWAY_FNAME")
+            fh = fopen(gateway_fname,"a", 1)
+            for(i=1;i<=nclusters;i++) {
+                tmpdir_i = tmpdir+"__pll"+parallelid+"_tmpdir"+strofreal(i, "%04.0f")
+                mkdir(tmpdir_i,1) // fput(fh, "mkdir "+c("tmpdir")+"/"+parallelid+strofreal(i,"%04.0f"))
+                fput(fh, `"export TEMP=""'+tmpdir_i+`"""')
+                fput(fh, paralleldir+`" -e -q do ""'+pwd()+"__pll"+parallelid+"_do"+strofreal(i,"%04.0f")+`".do" &"')
+            }
+            fclose(fh)
         }
-        
-        fput(fh, "exit")
-        
-        fclose(fh)
-        
-        stata("winexec __pll"+parallelid+"_shell.bat")
+        else{
+            unlink("__pll"+parallelid+"_shell.bat")
+            fh = fopen("__pll"+parallelid+"_shell.bat","w", 1)
+            
+            // Writing file
+            for(i=1;i<=nclusters;i++) {
+                mkdir(tmpdir+"__pll"+parallelid+"_tmpdir"+strofreal(i, "%04.0f"),1)
+                fwrite(fh, "start /MIN /HIGH set TEMP="+tmpdir+"__pll"+parallelid+"_tmpdir"+strofreal(i,"%04.0f")+" ^& ")
+                fput(fh, paralleldir+`" /e /q do ""'+pwd()+"__pll"+parallelid+"_do"+strofreal(i,"%04.0f")+`".do"^&exit"')
+            }
+            
+            fput(fh, "exit")
+            
+            fclose(fh)
+            
+            stata("winexec __pll"+parallelid+"_shell.bat")
+        }
     }
     
     /* Waits until each process ends */
@@ -1924,7 +1939,7 @@ mata:
 *!{dup 78:{c -}}{asis}
 real scalar parallel_setstatapath(string scalar statadir, | real scalar force) {
 
-    string scalar bit, flv
+    string scalar bit, flv, fname, sys_dir
 
     // Is it 64bits?
     if (c("osdtl") != "" | c("bit") == 64) bit = "-64"
@@ -1939,9 +1954,18 @@ real scalar parallel_setstatapath(string scalar statadir, | real scalar force) {
             else if (c("flavor") == "IC") flv = ""
         
             /* If the version is less than eleven */
-            if (c("stata_version") < 11) statadir = c("sysdir_stata")+"w"+flv+"Stata.exe"
-            else statadir = c("sysdir_stata")+"Stata"+flv+bit+".exe"
-        
+            if (c("stata_version") < 11) fname = "w"+flv+"Stata.exe"
+            else fname = "Stata"+flv+bit+".exe"
+            
+            statadir = c("sysdir_stata") + fname
+            
+            //might need to convert to cygwin path-name
+            if (c("mode")=="batch"){
+                if (!force) if (!fileexists(statadir)) return(601)
+                statadir = "/cygdrive/"+substr(c("sysdir_stata"),1,1)+"/"+substr(c("sysdir_stata"),4,.) + fname
+                force=1
+            }
+            
         }
         else if (regexm(c("os"), "^MacOS.*")) { // MACOS
